Added control pty for echolink module
This pty (ECHOLINK_PTY) let's one disconnect clients.
Sending the word 'KILL' will disconnect the current talker, if there is one.
Sending <CALL>:D will disconnect the client associated with the <CALL> call sign.
The purpose of this is to allow the operator to disconnect users that forget to un-key their transmit.