Added a control pty for the voter
This let's the operator control the connected RXes, ie. enable/disable them.
For this, RXes get a new flag m_enabled, get-able by isEnabled, set-able by setEnabled.
The Voter has been changed on multiple locations in the code to accomodate this.
The state_pty now has an extra character marking for the disabled state: '#'
The m_sql_open_hidden was put in to 'follow' the squelch state while in disabled state.
 This allows the code to set the correct sqluelch state on re-enable.